bind.sh: rearrange directory structure
Move ptex-sys/ up into the top-level. It is a build artifact
but it is convenient to have it here.

The .gitignore pattern is to allow us to have ptex/
ptex-sys/ and ptex-rs/ adjacent to themselves at the
top-level. The ptex/ location is not required, as we
look at ${PTEX_ROOT}, but the .gitignore allows it in-tree.

Signed-off-by: David Aguilar <davvid@gmail.com>