ptex-sys.h: remove unnecessary pointers checks and const casts
Let the C++ bridge assume that it's always going to be called with valid
pointers. The only nullptr checks that remain are the functions called
by drop(), i.e. the destruction functions.

Expose the mutability of the underlying C++ library into the sys
signatures so that we can stop using const_cast.