types: replace TryFrom with From implementations
We can eliminate one dimension of errors by defining what happens when
the core library returns unknown values.

The enums no longer use `repr(u32)` so that rust can use the
minimumally-sized representation for each enum. Any internal conversion
to the Ptex u32/i32 types are minimal and not a performance concern.
This improves the documentation as well since the repr values is no
longer displayed.